refactor parsing execution modes from config
this allows to implement custom execution modes classes in modules.
