+++ sdl-core/src/main/scala/io/smartdatalake/config/ConfigImplicits.scala
@@ -25,6 +25,7 @@ import io.smartdatalake.util.hdfs.SparkRepartitionDef
 import io.smartdatalake.util.misc.SchemaProviderType.SchemaProviderType
 import io.smartdatalake.util.misc.SchemaUtil
 import io.smartdatalake.util.secrets.SecretProviderConfig
+import io.smartdatalake.workflow.action.executionMode.ExecutionMode
 import io.smartdatalake.workflow.action.generic.transformer.{GenericDfTransformer, GenericDfsTransformer}
 import io.smartdatalake.workflow.action.script.ParsableScriptDef
 import io.smartdatalake.workflow.action.spark.customlogic._
@@ -92,7 +93,6 @@ trait ConfigImplicits {
   implicit val sparkUdfCreatorConfigReader: ConfigReader[SparkUDFCreatorConfig] = ConfigReader.derive[SparkUDFCreatorConfig]
   implicit val sparkRepartitionDefReader: ConfigReader[SparkRepartitionDef] = ConfigReader.derive[SparkRepartitionDef]
   implicit val secretProviderConfigReader: ConfigReader[SecretProviderConfig] = ConfigReader.derive[SecretProviderConfig]
-  implicit val executionModeReader: ConfigReader[ExecutionMode] = ConfigReader.derive[ExecutionMode]
   implicit val conditionReader: ConfigReader[Condition] = ConfigReader.derive[Condition]
   implicit val authModeReader: ConfigReader[AuthMode] = ConfigReader.derive[AuthMode]
   implicit val saveModeOptionsReader: ConfigReader[SaveModeOptions] = ConfigReader.derive[SaveModeOptions]
@@ -151,5 +151,14 @@ trait ConfigImplicits {
   implicit val expectationReader: ConfigReader[Expectation] = ConfigReader.fromTry { (c, p) =>
     implicit val instanceRegistry: InstanceRegistry = Environment._instanceRegistry
     ConfigParser.parseConfigObject[Expectation](c.getConfig(p))
+  }
+
+  /**
+   * A reader that reads [[ExecutionMode]] values.
+   * Note that Expectation must be parsed according to it's 'type' attribute by using SDL ConfigParser.
+   */
+  implicit val executionModeReader: ConfigReader[ExecutionMode] = ConfigReader.fromTry { (c, p) =>
+    implicit val instanceRegistry: InstanceRegistry = Environment._instanceRegistry
+    ConfigParser.parseConfigObject[ExecutionMode](c.getConfig(p))
   }
 }
