+++ sdl-core/src/main/scala/io/smartdatalake/definitions/Condition.scala
@@ -22,6 +22,7 @@ package io.smartdatalake.definitions
 import io.smartdatalake.config.ConfigurationException
 import io.smartdatalake.config.SdlConfigObject.ConfigObjectId
 import io.smartdatalake.util.spark.SparkExpressionUtil
+import io.smartdatalake.workflow.action.executionMode.PartitionDiffMode
 
 import scala.reflect.runtime.universe.TypeTag
