+++ sdl-core/src/main/scala/io/smartdatalake/workflow/InitSubFeed.scala
@@ -20,8 +20,8 @@
 package io.smartdatalake.workflow
 
 import io.smartdatalake.config.SdlConfigObject.DataObjectId
-import io.smartdatalake.definitions.ExecutionModeResult
 import io.smartdatalake.util.hdfs.PartitionValues
+import io.smartdatalake.workflow.action.executionMode.ExecutionModeResult
 import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 /**
