+++ sdl-core/src/main/scala/io/smartdatalake/workflow/ScriptSubFeed.scala
@@ -20,9 +20,9 @@
 package io.smartdatalake.workflow
 
 import io.smartdatalake.config.SdlConfigObject.DataObjectId
-import io.smartdatalake.definitions.ExecutionModeResult
 import io.smartdatalake.util.hdfs.PartitionValues
 import io.smartdatalake.util.misc.ScalaUtil.optionalizeMap
+import io.smartdatalake.workflow.action.executionMode.ExecutionModeResult
 
 /**
  * A ScriptSubFeed is used to notify DataObjects and subsequent actions about the completion of a script.
