+++ sdl-core/src/main/scala/io/smartdatalake/workflow/SubFeed.scala
@@ -19,10 +19,10 @@
 package io.smartdatalake.workflow
 
 import io.smartdatalake.config.SdlConfigObject.DataObjectId
-import io.smartdatalake.definitions.ExecutionModeResult
 import io.smartdatalake.util.dag.DAGResult
 import io.smartdatalake.util.hdfs.PartitionValues
 import io.smartdatalake.util.misc.SmartDataLakeLogger
+import io.smartdatalake.workflow.action.executionMode.ExecutionModeResult
 
 /**
  * A SubFeed transports references to data between Actions.
