+++ sdl-core/src/test/scala/io/smartdatalake/util/misc/SparkExpressionUtilTest.scala
@@ -21,9 +21,9 @@ package io.smartdatalake.util.misc
 
 import io.smartdatalake.config.InstanceRegistry
 import io.smartdatalake.config.SdlConfigObject.DataObjectId
-import io.smartdatalake.definitions.DefaultExecutionModeExpressionData
 import io.smartdatalake.testutils.TestUtil
 import io.smartdatalake.util.spark.SparkExpressionUtil
+import io.smartdatalake.workflow.action.executionMode.DefaultExecutionModeExpressionData
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.custom.ExpressionEvaluator
 import org.apache.spark.sql.functions.udf
