+++ sdl-lang/src/main/scala/io/smartdatalake/meta/GenericTypeUtil.scala
@@ -1,8 +1,9 @@
 package io.smartdatalake.meta
 
 import com.github.takezoe.scaladoc.{Scaladoc => ScaladocAnnotation}
-import io.smartdatalake.definitions.{AuthMode, ExecutionMode, SaveModeOptions}
+import io.smartdatalake.definitions.{AuthMode, SaveModeOptions}
 import io.smartdatalake.util.misc.SmartDataLakeLogger
+import io.smartdatalake.workflow.action.executionMode.ExecutionMode
 import io.smartdatalake.workflow.action.generic.transformer.{GenericDfTransformer, GenericDfsTransformer, ValidationRule}
 import io.smartdatalake.workflow.action.script.ParsableScriptDef
 import io.smartdatalake.workflow.action.{Action, ActionMetadata}
